server/requirements.txt
fastapi
uvicorn
psycopg2-binary
requests
python-jose[cryptography]
passlib[bcrypt]
python-multipart